Skip points component uninstall on large networks
We skip the per-site uninstallation of the points component when
uninstalling on a large network.

See #68

+++ src/components/points/uninstall.php
@@ -45,46 +45,50 @@ if ( is_multisite() ) {
 	delete_site_option( 'wordpoints_default_points_type' );
 	delete_site_option( 'wordpoints_points_types_hooks' );
 
-	$blog_ids = $wpdb->get_col( "SELECT blog_id FROM {$wpdb->blogs}" );
+	// On large networks we don't attempt the per-site uninstall.
+	if ( ! wp_is_large_network() ) {
 
-	$original_blog_id = get_current_blog_id();
+		$blog_ids = $wpdb->get_col( "SELECT blog_id FROM {$wpdb->blogs}" );
 
-	foreach ( $blog_ids as $blog_id ) {
+		$original_blog_id = get_current_blog_id();
 
-		switch_to_blog( $blog_id );
+		foreach ( $blog_ids as $blog_id ) {
 
-		$points_types = wordpoints_get_points_types();
+			switch_to_blog( $blog_id );
 
-		foreach ( $points_types as $slug => $settings ) {
+			$points_types = wordpoints_get_points_types();
 
-			delete_metadata( 'comment', 0, "wordpoints_last_status-{$slug}", '', true );
+			foreach ( $points_types as $slug => $settings ) {
 
-			$prefix = $wpdb->get_blog_prefix();
-			delete_metadata( 'user', 0, $prefix . "wordpoints_points-{$slug}", '', true );
-			delete_metadata( 'user', 0, $prefix . 'wordpoints_points_period_start', '', true );
+				delete_metadata( 'comment', 0, "wordpoints_last_status-{$slug}", '', true );
+
+				$prefix = $wpdb->get_blog_prefix();
+				delete_metadata( 'user', 0, $prefix . "wordpoints_points-{$slug}", '', true );
+				delete_metadata( 'user', 0, $prefix . 'wordpoints_points_period_start', '', true );
+			}
+
+			delete_option( 'wordpoints_points_types' );
+			delete_option( 'wordpoints_default_points_type' );
+			delete_option( 'wordpoints_points_types_hooks' );
+
+			delete_option( 'wordpoints_hook-wordpoints_registration_points_hook' );
+			delete_option( 'wordpoints_hook-wordpoints_post_points_hook' );
+			delete_option( 'wordpoints_hook-wordpoints_comment_points_hook' );
+			delete_option( 'wordpoints_hook-wordpoints_periodic_points_hook' );
+
+			delete_option( 'widget_wordpoints_points_logs_widget' );
+			delete_option( 'widget_wordpoints_top_users_widget' );
+			delete_option( 'widget_wordpoints_points_widget' );
+
+			wordpoints_remove_custom_caps( array_keys( wordpoints_points_get_custom_caps() ) );
 		}
 
-		delete_option( 'wordpoints_points_types' );
-		delete_option( 'wordpoints_default_points_type' );
-		delete_option( 'wordpoints_points_types_hooks' );
+		switch_to_blog( $original_blog_id );
 
-		delete_option( 'wordpoints_hook-wordpoints_registration_points_hook' );
-		delete_option( 'wordpoints_hook-wordpoints_post_points_hook' );
-		delete_option( 'wordpoints_hook-wordpoints_comment_points_hook' );
-		delete_option( 'wordpoints_hook-wordpoints_periodic_points_hook' );
-
-		delete_option( 'widget_wordpoints_points_logs_widget' );
-		delete_option( 'widget_wordpoints_top_users_widget' );
-		delete_option( 'widget_wordpoints_points_widget' );
-
-		wordpoints_remove_custom_caps( array_keys( wordpoints_points_get_custom_caps() ) );
+		// See http://wordpress.stackexchange.com/a/89114/27757
+		unset( $GLOBALS['_wp_switched_stack'] );
+		$GLOBALS['switched'] = false;
 	}
-
-	switch_to_blog( $original_blog_id );
-
-	// See http://wordpress.stackexchange.com/a/89114/27757
-	unset( $GLOBALS['_wp_switched_stack'] );
-	$GLOBALS['switched'] = false;
 
 } else {
